Move some hooks from widget to the features folder
