+++ src/features/PassSurvey/hooks/index.ts
@@ -1,3 +1,6 @@
+export * from './useSurveyDataQuery';
 export * from './useStartSurvey';
 export * from './useSummaryData';
 export * from './useItemTimer';
+export * from './useAnswers';
+export * from './useSubmitAnswersMutation';
