+++ src/features/PassSurvey/hooks/useAnswers.ts
@@ -2,12 +2,12 @@ import { useCallback, useContext } from 'react';
 
 import { v4 as uuidV4 } from 'uuid';
 
-import { SurveyContext } from '../../lib';
-import { generateUserPublicKey } from '../generateUserPublicKey';
-import { getFirstResponseDataIdentifierTextItem } from '../getFirstResponseDataIdentifierTextItem';
-import { getScheduledTimeFromEvents } from '../getScheduledTimeFromEvents';
-import { mapAlerts, mapToAnswers } from '../mappers';
-import { prepareItemAnswers } from '../prepareItemAnswers';
+import { SurveyContext } from '../../../widgets/Survey/lib';
+import { generateUserPublicKey } from '../../../widgets/Survey/model/generateUserPublicKey';
+import { getFirstResponseDataIdentifierTextItem } from '../../../widgets/Survey/model/getFirstResponseDataIdentifierTextItem';
+import { getScheduledTimeFromEvents } from '../../../widgets/Survey/model/getScheduledTimeFromEvents';
+import { mapAlerts, mapToAnswers } from '../../../widgets/Survey/model/mappers';
+import { prepareItemAnswers } from '../../../widgets/Survey/model/prepareItemAnswers';
 
 import { ActivityPipelineType, GroupProgress } from '~/abstract/lib';
 import { useEncryptPayload } from '~/entities/activity';
