+++ src/widgets/Survey/model/hooks/index.ts
@@ -1,5 +1,2 @@
-export * from './useSurveyDataQuery';
-export * from './useSubmitAnswersMutation';
-export * from './useAnswers';
-export * from './useSurvey';
+export * from './useSurveyState';
 export * from './useAutoForward';
