+++ src/widgets/Survey/model/hooks/useSurveyState.ts
